(#26) Добавить механизм DB-миграций (#27)
* (#26) Добавить механизм DB-миграций

Тип всех id изменён с int на long.
Подключен flyway.
Написана базовая миграция.
Включен "ddl-auto=validate" и исправлена ошибка "Schema-validation: missing table [hibernate_sequences]".

+++ backend/src/main/java/com/github/viktor235/rggassistant/models/entitys/modifiers/CollectedItem.java
@@ -15,7 +15,12 @@ public class CollectedItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
+
+//    @Schema(description = "Date the item was added to the collection")
+//    @Column(name = "collection_date", insertable = false)
+////    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private ZonedDateTime collectionDate;
 
     @JoinColumn(name = "item_id", referencedColumnName = "id")
     @ManyToOne
