+++ backend/src/main/java/com/github/viktor235/rggassistant/repositories/CollectedGameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollectedGameRepository extends JpaRepository<CollectedGamePlatform, Integer> {
+public interface CollectedGameRepository extends JpaRepository<CollectedGamePlatform, Long> {
 }
