+++ backend/src/main/java/com/github/viktor235/rggassistant/repositories/PlatformRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlatformRepository extends JpaRepository<Platform, Integer> {
+public interface PlatformRepository extends JpaRepository<Platform, Long> {
 }
